Fixed Fate generation.,
for() procs iwth a single variable in the () become loops. You want to create a break here instead of a return to end the loop, otherwise it will continue to loop back in on itself.
As well the line where party/p being changed to usr instead of world fixes the isues were all parties IDs with all player IDs were compared. The in usr changes it so only the aprty object in the user compares it's IDs to mobs in the world.
For() procs are dangerous. Try to find ways around them or impliment breaks within them in the future.